Revert CLI params
